+++ changelog.md
@@ -2,16 +2,22 @@ Changelog for Laravel Models Generator
 ======================================
 
 3.0.0-dev
----------
+------------------
 
+- Ajout des tables "cache" et "sessions" dans l'option de config "ignored_tables".
 - Ajout de la possibilité d'ignorer des relations (option de config "ignored_relations").
+- Ajout de l'option de commande "--table" ("-t") pour spécifier les tables à générer.
+- Ajout de l'option de commande "--update" ("-u") pour mettre à jour les modèles existants.
+- Modification du format des options de config "polymorphic_relations" et "one_to_one_relations".
 - Utilisation des noms de tables au lieu des noms de modèles pour les nommages au pluriel.
 - Utilisation du mot-clé "Via" à la place de "Of" pour les précisions des relations.
 - Ajout du préfixe "pivot" aux noms des relations "has many" vers les tables pivots.
 - Suppression de l'extention d'Eloquent dans le template du modèle.
 - Suppression de l'appel à la méthode "withTimestamps" dans le template de la relation "belongsToMany".
+- Gestion des cas de doublons dans les noms de relations.
 - Ordre des méthodes des relations par nom de méthode et non par nom de table.
 - Utilisation de l'IoC pour instancier le driver.
+- Réécriture des commentaires dans la config.
 
 2.0.3 (2016-03-22)
 ------------------